refactor: add region utilities for clipboard data handling and offset calculations
