+++ src/utilities/index.ts
@@ -10,6 +10,7 @@ export * from "./math/math";
 export * from "./math2d/math2d";
 export * from "./parsing/parsing";
 export * from "./Processed/Processed";
+export * from "./region/region";
 export * from "./sorting/sorting";
 export * from "./table/table";
 export * from "./templates/templates";
